+++ packages/terra-date-picker/src/react-datepicker/calendar.jsx
@@ -245,9 +245,9 @@ export default class Calendar extends React.Component {
 
     this.todayBtnRef = React.createRef();
     this.closeBtnRef = React.createRef();
+    this.previousMonthBtnRef = React.createRef();
+    this.nextMonthBtnRef = React.createRef();
     this.monthRef;
-    this.previousMonthBtnRef;
-    this.nextMonthBtnRef;
     this.monthDropdownRef;
     this.yearDropdownRef;
   }
@@ -327,14 +327,6 @@ export default class Calendar extends React.Component {
 
   handleMonthBlur = () => {
     this.setState({ calendarIsKeyboardFocused: false })
-  }
-
-  setPreviousMonthBtnRef = (node) => {
-    this.previousMonthBtnRef = node;
-  }
-
-  setNextMonthBtnRef = (node) => {
-    this.nextMonthBtnRef = node;
   }
 
   setMonthRef = (node) => {
@@ -449,15 +441,16 @@ export default class Calendar extends React.Component {
     return (
       <FormattedMessage id="Terra.datePicker.previousMonth">
         {text => (
-          <Button
+          <button
+            type="button"
             className={cx('react-datepicker-navigation--previous')}
-            icon={<span className={cx('prev-month-icon')} />}
-            isIconOnly
-            variant="utility"
-            text={text}
+            aria-label={text}
             onClick={this.decreaseMonth}
             onKeyDown={this.handlePreviousMonthBtnKeyDown}
-            refCallback={this.setPreviousMonthBtnRef} />
+            ref={this.previousMonthBtnRef}
+          >
+            <span className={cx('prev-month-icon')} />
+          </button>
         )}
       </FormattedMessage>
     )
@@ -471,14 +464,15 @@ export default class Calendar extends React.Component {
     return (
       <FormattedMessage id="Terra.datePicker.nextMonth">
         {text => (
-          <Button
+          <button
+            type="button"
             className={cx('react-datepicker-navigation--next')}
-            icon={<span className={cx('next-month-icon')} />}
-            isIconOnly
-            variant="utility"
-            text={text}
+            aria-label={text}
             onClick={this.increaseMonth}
-            refCallback={this.setNextMonthBtnRef} />
+            ref={this.nextMonthBtnRef}
+          >
+            <span className={cx('next-month-icon')} />
+          </button>
         )}
       </FormattedMessage>
     )